ProyectoDeEspecialidad/2021-1-Grupo12-Backend/osam_backend/backend_app/db_serializers/enterprise_serializer.py
from backend_app.db_models.enterprise import Enterprise
from rest_framework import serializers


class EnterpriseSerializer(serializers.ModelSerializer):
    """Enterprise serializer."""

    class Meta:
        """Meta."""

        model = Enterprise
        fields = (
            "id",
            "name"
        )
