ProyectoDeEspecialidad/2021-1-Grupo12-Backend/pyproject.toml
[tool.poetry]
name = "2021-1-grupo12-backend"
version = "0.1.0"
description = ""
authors = ["AbyssalBit <hrhartmann@uc.cl>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.0.5"
djangorestframework = "^3.12.4"
djongo = "^1.3.4"
dnspython = "^2.1.0"
pre-commit = "^2.12.1"
pymongo = "^3.11.3"
bcrypt = "^3.2.0"
gunicorn = "^20.1.0"
django-cors-headers = "^3.7.0"
python-dotenv = "^0.17.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
